// apps/web/src/api/generated/model/dailyUsage.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Lighthouse API
 * API for the Lighthouse project
 * OpenAPI spec version: 1.0.0
 */

export interface DailyUsage {
  date: string
  promptTokens: number
  completionTokens: number
  totalTokens: number
  /** Nombre de messages */
  count: number
}