// apps/web/src/api/generated/model/usageByIntent.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Lighthouse API
 * API for the Lighthouse project
 * OpenAPI spec version: 1.0.0
 */
import type { UsageByIntentIntent } from "./usageByIntentIntent"

export interface UsageByIntent {
  intent: UsageByIntentIntent
  totalTokens: number
  count: number
}